+++ tests/PhpSpreadsheetTests/Reader/HtmlTest.php
@@ -329,6 +329,48 @@ class HtmlTest extends TestCase
         $this->assertContains("\n", $cellValue);
     }
 
+    public function testCanLoadFromStringIntoExistingSpreadsheet()
+    {
+        $html = '<table>
+                    <tr>
+                        <td>Hello World</td>
+                    </tr>
+                    <tr>
+                        <td>Hello<br />World</td>
+                    </tr>
+                    <tr>
+                        <td>Hello<br>World</td>
+                    </tr>
+                </table>';
+        $reader = new Html();
+        $spreadsheet = $reader->loadFromString($html);
+        $firstSheet = $spreadsheet->getSheet(0);
+
+        $cellStyle = $firstSheet->getStyle('A1');
+        self::assertFalse($cellStyle->getAlignment()->getWrapText());
+
+        $cellStyle = $firstSheet->getStyle('A2');
+        self::assertTrue($cellStyle->getAlignment()->getWrapText());
+        $cellValue = $firstSheet->getCell('A2')->getValue();
+        $this->assertContains("\n", $cellValue);
+
+        $cellStyle = $firstSheet->getStyle('A3');
+        self::assertTrue($cellStyle->getAlignment()->getWrapText());
+        $cellValue = $firstSheet->getCell('A3')->getValue();
+        $this->assertContains("\n", $cellValue);
+
+        $reader->setSheetIndex(1);
+        $html = '<table>
+                    <tr>
+                        <td>Goodbye World</td>
+                    </tr>
+                </table>';
+
+        self::assertEquals(1, $spreadsheet->getSheetCount());
+        $spreadsheet = $reader->loadFromString($html, $spreadsheet);
+        self::assertEquals(2, $spreadsheet->getSheetCount());
+    }
+
     /**
      * @param string $html
      *